3 Layer Arch - moving controller logic out of server handlers
