+++ backend/src/controllers/other.controller.ts
@@ -1,0 +1,18 @@
+import { Request, Response } from 'express';
+import { clear, echo } from '../utils/other';
+
+export const getEcho = (req: Request, res: Response) => {
+  const result = echo(req.query.echo as string);
+  if ('error' in result) {
+    res.status(400);
+  }
+  return res.json(result);
+};
+
+export const getClear = (req: Request, res: Response) => {
+  const result = clear();
+  if ('error' in result) {
+    return res.status(400).json(result);
+  }
+  res.json(result);
+};
